requirements.txt
langchain
langgraph
langchain-pinecone
boto3
pillow
flask
pdf2image
reportlab
nose2